♻️ refactor: RestaurantService 에 있던 영업 시간 조회 부분 이동
- OperatingHourService 로 영업 시간 조회 로직 이동
